Move rendering code into useCesiumViewer.
