+++ src/components/lava-coder/lava-coder-view.tsx
@@ -1,14 +1,9 @@
-import type { ImageryLayer } from "cesium";
-import { useEffect, useRef, useState } from "react";
-import { autorun } from "mobx";
+import { useEffect, useState } from "react";
 import { observer } from "mobx-react-lite";
-import { lavaElevations, lavaSimulation } from "../../stores/lava-simulation-store";
+import { lavaSimulation } from "../../stores/lava-simulation-store";
 import IconButton from "../buttons/icon-button";
 import RasterWorker from "./raster.worker";
 import { useCesiumViewer } from "./use-cesium-viewer";
-import { visualizeLava } from "./visualize-lava";
-
-const Cesium = (window as any).Cesium;
 
 import "./lava-coder-view.css";
 
@@ -21,11 +16,8 @@ interface IProps {
 export const LavaCoderView = observer(function LavaCoderView({ width, height, margin }: IProps) {
   const [lavaCoderElt, setLavaCoderElt] = useState<HTMLDivElement | null>(null);
   const [showHazardZones, setShowHazardZones] = useState(false);
-  const lavaLayerRef = useRef<ImageryLayer | null>(null);
-  const oldLavaLayerRef = useRef<ImageryLayer | null>(null);
-  const { coveredCells, raster } = lavaSimulation;
 
-  const { hazardZones, viewer } = useCesiumViewer(lavaCoderElt);
+  const { hazardZones } = useCesiumViewer(lavaCoderElt);
 
   // Load the elevation data
   useEffect(() => {
@@ -50,30 +42,6 @@ export const LavaCoderView = observer(function LavaCoderView({ width, height, ma
       rasterWorker.terminate();
     };
   }, []);
-
-  // Update the lava layer
-  useEffect(() => {
-    return autorun(() => {
-      if (!coveredCells || !lavaElevations || !raster || !viewer) return;
-
-      const oldLayer = oldLavaLayerRef.current;
-      oldLavaLayerRef.current = lavaLayerRef.current;
-
-      const url = visualizeLava(raster, lavaElevations);
-      lavaLayerRef.current = Cesium.ImageryLayer.fromProviderAsync(
-        Cesium.SingleTileImageryProvider.fromUrl(url, {
-          rectangle: Cesium.Rectangle.fromDegrees(
-            -155.673766,
-            19.370473,
-            -155.008440,
-            19.819655
-          )
-        })
-      );
-      if (lavaLayerRef.current) viewer.imageryLayers.add(lavaLayerRef.current);
-      if (oldLayer) viewer.imageryLayers.remove(oldLayer, true);
-    });
-  }, [coveredCells, raster, viewer]);
 
   function toggleHazardZones() {
     const newShowHazardZones = !showHazardZones;
