--- src/components/lava-coder/lava.scss
@@ -1,4 +1,0 @@
-.lava-output {
-  background-color: white;
-  padding: 10px;
-}
